Разбирает параметры url
