+++ react-gb/src/components/ChatsList.js
@@ -5,22 +5,27 @@ import InboxIcon from "@mui/icons-material/Inbox";
 import ListItemText from "@mui/material/ListItemText";
 import List from "@mui/material/List";
 import React from "react";
+import {Link, Outlet} from "react-router-dom";
 
-export const ChatList = ({chats}) => {
+export const ChatsList = ({chats}) => {
     return (
-        <List>
-            {chats.map(chat => {
-                return (
+        <div style={{display: "flex"}}>
+            <List>
+                {chats.map(chat => (
                     <ListItem disablePadding key={chat.id}>
-                        <ListItemButton>
-                            <ListItemIcon>
-                                <InboxIcon />
-                            </ListItemIcon>
-                            <ListItemText primary={chat.name} />
-                        </ListItemButton>
+                        <Link to={`/chats/${chat.id}`}>
+                            <ListItemButton>
+                                <ListItemIcon>
+                                    <InboxIcon/>
+                                </ListItemIcon>
+                                <ListItemText primary={chat.name}/>
+                            </ListItemButton>
+                        </Link>
                     </ListItem>
-                )
-            })}
-        </List>
+                ))}
+            </List>
+            <Outlet/>
+        </div>
+
     )
 }
